Resolved warning for comparing integers of different signedness

+++ src/layers/learning/channelwise_fully_connected.cpp
@@ -41,7 +41,7 @@ void channelwise_fully_connected_layer<T,L,D>::write_specific_proto(lbann_data::
   proto.set_datatype(proto::ProtoDataType<T>);
   auto* msg = proto.mutable_channelwise_fully_connected();
   auto const& dims = this->get_output_dims();
-  for (int ii = 1; ii < dims.size(); ii++)
+  for (size_t ii = 1; ii < dims.size(); ii++)
     msg->add_output_channel_dims(dims[ii]);
   msg->mutable_bias()->set_value(m_has_bias);
   msg->mutable_transpose()->set_value(m_transpose);
